Changed new ticket screen header to fit heuristic analysis
Moved all buttons to the top, put the back button on the left and changed it to a black colour to make it stand out from the restart button as it's not the same severity.

resetart button has been placed in the middle and renamed to 'start over' for clearer instructions. Kept red for the severity.

'confirm' has been switch to 'next' and renamed to 'next'

Footer is empty as a placeholder for keeping track of progress in the application.

This is a template for all other screens