Updated back & forward buttons to keep track of screen traversal order
Each screen is given an index (number as a string), it is appended to a string, a switch board translates this number to the right url.

It's a string because passing data through the url is processed as a string.

it's possible to use .js export and import but we would need to set up a webserver for this to work.  does not work with file:// as we are doing now

This string through the url method works with file://